#jka_606ロジック修正

+++ app/Http/Controllers/Api/Manager/ChapterController.php
@@ -23,13 +23,14 @@ class ChapterController extends Controller
     public function delete(Request $request)
     {
         $instructorId = Auth::guard('instructor')->user()->id;
+        // 配下の講師情報を取得
         $manager = Instructor::with('managings')->find($instructorId);
         $instructorIds = $manager->managings->pluck('id')->toArray();
         $instructorIds[] = $instructorId;
 
-        //自分と配下のインストラクターが持っている講座を取得
         $course = Course::FindOrFail($request->course_id);
-
+        $chapter = Chapter::with('course')->findOrFail($request->chapter_id);
+        
         if (!in_array($course->instructor_id, $instructorIds, true)) {
             // 自分、または配下の講師の講座のチャプターでなければエラー応答
             return response()->json([
@@ -37,9 +38,13 @@ class ChapterController extends Controller
                 'message' => "Forbidden, not allowed to delete this chapter.",
             ], 403);
         }
-
-        $chapter = Chapter::with('course')->findOrFail($request->chapter_id);
-        
+        if ((int) $request->course_id !== $chapter->course->id) {
+            // 指定したコースに属するチャプターでなければエラー応答
+            return response()->json([
+                'result'  => false,
+                'message' => 'invalid course_id.',
+            ], 403);
+        }
         $chapter->delete();
         return response()->json([
             "result" => true
